wolfram-language-server/pyproject.toml
[project]
name = "wolfram-language-server"
version = "0.1.0"
description = "Backend API server for executing Wolfram Language scripts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "wolframclient>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wolfram_language_server"]
